refactor(mobile): Shrine Knowledge Fact Section component境界(PR-M3A)
PR-M2でShrine Detail screen内部へ直接実装したKnowledge Fact UI
（「祭神・由緒」セクション）を、apps/mobile/components/shrine/
ShrineKnowledgeFactSection.tsxへ責務分離する。挙動変更ではない。

- ShrineKnowledgeFactSection: ShrineFactViewModelのみをpropsとして
  受け取る表示専用component。API fetch・toShrine・Evidence判定・
  Recommendation・Navigation・Analytics・Storage・Favorites・Authの
  いずれにも依存しない。これにより将来のMobile Evidence Flow
  integration testを、Detail screen全体の大量mock無しで実施できる
  境界を用意する。
- apps/mobile/app/shrines/[id].tsx: buildShrineFactViewModel()の
  結果を<ShrineKnowledgeFactSection factViewModel={factViewModel} />
  へ渡すだけの構造に変更。重複していたJSX・StyleSheetエントリ・
  history_typeラベル対応表・disputed文言定数を削除（screenは186行の
  純減）。Legacy「神社について」・表示順は変更していない。

Manual QA（Expo web、375px幅）で、full/disputed表示・複数祭神・長い
祭神名の折り返し（flexShrink/flexWrap/maxWidth維持）・複数History・
Knowledge Fact 0件時の非表示・Legacy「神社について」共存が、PR-M2と
完全に同一の見た目であることを確認した。QA用の一時コードは検証後に
削除済み（PR diffに含まれない）。

既存テスト（apps/mobile/lib/__tests__/shrineKnowledgeFact.test.ts、
18 tests）は無変更のまま全PASS。新規RN component testing library・
新規テストは追加していない。Backend・Web・docs・Model・Migration・
API Schema・Recommendation・Evidence Gate・Navigation・vitest config・
CI workflow・package.json・lockfileは変更しない。
